requirements.txt
chromadb
newsapi-python
duckduckgo-search
schedule
edge_tts
pygame
SpeechRecognition
pyaudio